backend-src/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.42
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
requests==2.28.1
Flask-CORS==3.0.10
openai==1.9.0
python-dotenv==1.0.0
